# docs/openapi.yaml
openapi: 3.0.3
info:
  title: IhaleTakip API
  version: 1.0.0
  description: Kamu ihalesi takip uygulaması backend servisi
paths:
...
  /api/v1/ekap/tenders/:
    get:
      operationId: api_v1_ekap_tenders_list
      description: 'İhaleleri **kendi veritabanımızdan** arar — EKAP''a istek gitmez,
        rate limit yoktur. Alan adları EKAP''ın kendi adlandırmasıyla döner.


        Yanıt: `data.list` (ihaleler), `data.totalCount` (toplam kayıt), `data.page`.


        **Filtre parametre adları = `Tender` model alan adlarıdır** (tek adlandırma).
        Liste alanları (`il_id`, `ihale_tip`, `ihale_usul`, `ihale_durum`, `yasa_kapsami`,
        `idare_id`, `okas_kod`, `okas_adi`, `ozellik`) **virgülle ayrılmış** verilir
        (ör. `ihale_tip=1,3`).


        **İhale türü (`ihale_tip`):** 1 Mal Alımı · 2 Yapım · 3 Hizmet · 4 Danışmanlık


        **İhale usulü (`ihale_usul`):** 1 Açık İhale · 2 Belli İstekliler Arasında
        · 3 Pazarlık (MD 21 F) · 4 Doğrudan Temin


        **İhale durumu (`ihale_durum`):** 1 Taslak · 2/3 Katılıma Açık · 4 Değerlendirme
        Tamamlanmış · 5 Değerlendirmede · 6/10 İptal Edilmiş · 15 Sonuç İlanı Yayımlanmış
        · 20 Sözleşme İmzalanmış


        **İl id''leri (`il_id`)** EKAP''ın kendi il kodlarıdır (plaka değil!) — `GET
        /api/v1/ekap/cities/` ile alın. Örn. Ankara `251`, İstanbul `284`.


        Gelişmiş alanlar (`il_id`, `ihale_usul`, `yasa_kapsami`, `ozellik`, `okas_*`)
        detay senkronunda dolar; henüz detayı gelmemiş ihaleler bu filtrelerde eşleşmeyebilir.'
      summary: İhale ara / listele
      parameters:
      - in: query
        name: idare_adi
        schema:
          type: string
        description: Kurum (idare) adında geçen metin (kısmi eşleşme).
        examples:
          Kurum:
            value: belediyesi
      - in: query
        name: idare_detsis
        schema:
          type: string
        description: 'İdare DETSIS ağaç düğümü `detsis_no` listesi (virgülle). Üst
          düğüm (örn. bakanlık) seçilince altındaki tüm idarelerin ihaleleri gelir.
          Ağaç uçları: `GET /ekap/authorities/tree/`, `GET /ekap/authorities/search/`.'
        examples:
          BakanlıkDüğümü:
            value: '24308110'
            summary: Bakanlık düğümü
      - in: query
        name: idare_id
        schema:
          type: string
        description: İdare (kurum) id listesi (virgülle) — doğrudan `Tender.idare_id`
          ile eşleşir (yaprak seçim).
      - in: query
        name: ihale_adi
        schema:
          type: string
        description: Yalnızca ihale adında geçen metin (kısmi eşleşme).
        examples:
          MetinArama:
            value: bilgisayar
            summary: Metin arama
      - in: query
        name: ihale_durum
        schema:
          type: string
        description: İhale durumu id listesi.
        examples:
          KatılımaAçık:
            value: 2,3
            summary: Katılıma açık
      - in: query
        name: ihale_tarihi_max
        schema:
          type: string
        description: İhale tarihi üst sınırı. `GG.AA.YYYY`, `GG.AA.YYYY SS:dd` veya
          ISO-8601. Geçersiz tarih yok sayılır.
        examples:
          Tarih:
            value: 31.12.2026
      - in: query
        name: ihale_tarihi_min
        schema:
          type: string
        description: İhale tarihi alt sınırı. `GG.AA.YYYY`, `GG.AA.YYYY SS:dd` veya
          ISO-8601. Geçersiz tarih yok sayılır.
        examples:
          Tarih:
            value: 01.01.2026
      - in: query
        name: ihale_tip
        schema:
          type: string
        description: 'İhale türü id listesi: 1 Mal, 2 Yapım, 3 Hizmet, 4 Danışmanlık.'
        examples:
          Mal+Hizmet:
            value: 1,3
            summary: Mal + Hizmet
      - in: query
        name: ihale_usul
        schema:
          type: string
        description: İhale usulü id listesi (1-4).
        examples:
          AçıkIhale:
            value: '1'
            summary: Açık ihale
      - in: query
        name: ikn
        schema:
          type: string
        description: İKN'de geçen metin (kısmi eşleşme).
        examples:
          İKNParçası:
            value: 2026/271
            summary: İKN parçası
      - in: query
        name: ikn_sayi
        schema:
          type: string
        description: İKN sıra sayısı — İKN `.../SAYI` ile bitenler.
        examples:
          SıraNo:
            value: '271215'
            summary: Sıra no
      - in: query
        name: ikn_yil
        schema:
          type: string
        description: İKN yılı — İKN `YIL/...` ile başlayanlar.
        examples:
          2026Yılı:
            value: '2026'
            summary: 2026 yılı
      - in: query
        name: il_id
        schema:
          type: string
        description: İl id listesi (virgülle). `GET /ekap/cities/` ile alın.
        examples:
          Ankara+İstanbul:
            value: 251,284
            summary: Ankara + İstanbul
      - in: query
        name: ilan_tarihi_max
        schema:
          type: string
        description: İlan (yayın) tarihi üst sınırı. `GG.AA.YYYY`, `GG.AA.YYYY SS:dd`
          veya ISO-8601. Geçersiz tarih yok sayılır.
        examples:
          Tarih:
            value: 31.12.2026
      - in: query
        name: ilan_tarihi_min
        schema:
          type: string
        description: İlan (yayın) tarihi alt sınırı. `GG.AA.YYYY`, `GG.AA.YYYY SS:dd`
          veya ISO-8601. Geçersiz tarih yok sayılır.
        examples:
          Tarih:
            value: 01.01.2026
      - in: query
        name: okas_adi
        schema:
          type: string
        description: OKAS branş adı listesi (virgülle, kısmi eşleşme).
        examples:
          OKASAdı:
            value: yazılım
            summary: OKAS adı
      - in: query
        name: okas_kod
        schema:
          type: string
        description: OKAS branş kodu listesi (virgülle). Kod ile başlayan kalemleri
          olan ihaleler.
        examples:
          OKASKodu:
            value: '72415000'
            summary: OKAS kodu
      - in: query
        name: order
        schema:
          type: string
          enum:
          - ihale_tarihi
          - ilan_tarihi
          default: ihale_tarihi
        description: Sıralama alanı.
      - in: query
        name: ozellik
        schema:
          type: string
        description: 'Özellik anahtarları (virgülle): eIhale, kismiTeklifMi, altYukleniciCalistirilabilirMi,
          fiyatFarkiVerilecekMi, isDeneyimiGosterenBelgelerIsteniyorMu, meslekiTeknikYeterlilikBelgeleriIsteniyorMu,
          yabanciIsteklilereIzinVeriliyorMu, yerliIstekliyeFiyatAvantajiUgulaniyorMu,
          ekonomikVeMaliYeterlilikBelgeleriIsteniyorMu.'
      - in: query
        name: page
        schema:
          type: integer
          default: 1
        description: Sayfa numarası (1'den başlar).
      - in: query
        name: page_size
        schema:
          type: integer
          default: 10
        description: Sayfa boyutu. **En fazla 100** (aşan değer 100'e kırpılır).
      - in: query
        name: q
        schema:
          type: string
        description: Anahtar kelime — ihale adı + **kurum adı** + İKN içinde arar
          (mobil 'Anahtar Kelime' kutusu). Alan-özel arama için `ihale_adi`/`idare_adi`/`ikn`.
        examples:
          AnahtarKelime:
            value: siber güvenlik
            summary: Anahtar kelime
      - in: query
        name: siralamaTipi
        schema:
          type: string
          enum:
          - asc
          - desc
          default: desc
        description: Sıralama yönü.
      - in: query
        name: yasa_kapsami
        schema:
          type: string
        description: Yasa kapsamı id listesi (1=4734, 2=4734 dışı, 3=istisna). Detayı
          gelmemiş ihaleler (kapsam boş) dışlanmaz.
      tags:
      - ekap
      security:
      - jwtAuth: []
      - cookieAuth: []
      - {}
      responses:
        '200':
          description: No response body
...
  /api/v1/ekap/tenders/{key}/:
    get:
      operationId: api_v1_ekap_tenders_retrieve
      description: 'İhale detayını EKAP''ın ham detay şekliyle döner (`item` açılmış
        olarak).


        Detay veritabanında varsa oradan servis edilir; bayatsa arka planda yenilenir
        ve eldeki sürüm hemen döner. Detay hiç çekilmemişse EKAP''a **canlı** gidilir
        (bu istek yavaş olabilir) ve sonuç saklanır.


        EKAP''a ulaşılamaz ve elde eski bir kopya varsa `200` + `message: "Detay güncel
        değil."` döner. Hiç kayıt yoksa `404`.


        Yanıt `ilanList` alanını zaten içerir — ilanlar için ayrıca `/announcements/`
        çağırmanız gerekmez.'
      summary: İhale detayı
      parameters:
      - in: path
        name: key
        schema:
          type: string
        description: '**EKAP iç kimliği (`ekap_id`) kullanın** — ör. `1234567`.


          View, İKN ile de arama yapar; ancak bu rota `/` içeren bir değeri eşleştiremez
          ve `%2F` ile kodlamak da işe yaramaz (sunucu yolu çözerken geri `/` yapar).
          Dolayısıyla `2025/1234567` biçimindeki bir İKN burada **404 döner**; slash
          içermeyen `ekap_id` kullanın.'
        required: true
        examples:
          EKAPIçKimliği:
            value: '1234567'
            summary: EKAP iç kimliği
      tags:
      - ekap
      security:
      - jwtAuth: []
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
        '404':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT